+++ lib/crucible/safetensors/checksum.ex
@@ -1,6 +1,8 @@
 defmodule Crucible.Safetensors.Checksum do
   @moduledoc "Compatibility namespace for `CrucibleSafetensors.Checksum`."
 
-  defdelegate file_sha256(path), to: CrucibleSafetensors.Checksum
-  defdelegate file_sha256!(path), to: CrucibleSafetensors.Checksum
+  defdelegate file_sha256(path, opts \\ []), to: CrucibleSafetensors.Checksum
+  defdelegate file_sha256!(path, opts \\ []), to: CrucibleSafetensors.Checksum
+  defdelegate verify_file(path, expected, opts \\ []), to: CrucibleSafetensors.Checksum
+  defdelegate verify_file!(path, expected, opts \\ []), to: CrucibleSafetensors.Checksum
 end